[DOCUMENTATION] Pass over the code base

+++ Coil/Source/Coil/GUI/Components/GUITextInput.cpp
@@ -14,7 +14,7 @@ namespace Coil
 	{
 		CL_PROFILE_FUNCTION_MEDIUM()
 
-		ImGui::InputText(Properties.Label->CString(), InputBuffer->GetBuffer(), static_cast<int32>(InputBuffer->GetSize()) + 1, ImGuiInputTextFlags_CallbackResize, [](ImGuiInputTextCallbackData* data) -> int
+		ImGui::InputText(Properties.Label->CString(), InputBuffer->GetBuffer(), static_cast<size_t>(InputBuffer->GetSize()) + 1, ImGuiInputTextFlags_CallbackResize, [](ImGuiInputTextCallbackData* data) -> int
 		{
 			CL_PROFILE_FUNCTION_MEDIUM()
